combine maps and slices into collections
