+++ protocol/app/msgs/all_msgs_test.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/dydxprotocol/v4-chain/protocol/app/msgs"
 	"github.com/dydxprotocol/v4-chain/protocol/lib"
-	"github.com/dydxprotocol/v4-chain/protocol/lib/maps"
 	testapp "github.com/dydxprotocol/v4-chain/protocol/testutil/app"
 	"github.com/stretchr/testify/require"
 )
@@ -38,7 +37,7 @@ func TestAllTypeMessages(t *testing.T) {
 
 func TestAllTypeMessages_SumOfDistinctLists(t *testing.T) {
 	// The following should fail if there's a duplicate message in any of the lists.
-	expectedAllTypeMsgs := maps.MergeAllMapsMustHaveDistinctKeys(
+	expectedAllTypeMsgs := lib.MergeAllMapsMustHaveDistinctKeys(
 		msgs.AppInjectedMsgSamples,
 		msgs.InternalMsgSamplesAll,
 		msgs.NestedMsgSamples,
@@ -89,7 +88,7 @@ func TestAllTypeMessages_NoUnregisteredMsg(t *testing.T) {
 }
 
 func TestDisallowMsgs(t *testing.T) {
-	expectedDisallowMsgs := maps.MergeAllMapsMustHaveDistinctKeys(
+	expectedDisallowMsgs := lib.MergeAllMapsMustHaveDistinctKeys(
 		msgs.AppInjectedMsgSamples,
 		msgs.InternalMsgSamplesAll,
 		msgs.NestedMsgSamples,
@@ -103,7 +102,7 @@ func TestAllowMsgs(t *testing.T) {
 }
 
 func TestAllTypeMessages_SumOfAllowDisallow_MinusUnregistered(t *testing.T) {
-	expectedAllTypeMsgs := maps.MergeAllMapsMustHaveDistinctKeys(
+	expectedAllTypeMsgs := lib.MergeAllMapsMustHaveDistinctKeys(
 		msgs.DisallowMsgs,
 		msgs.AllowMsgs,
 	)
