+++ protocol/lib/ante/nested_msg_test.go
@@ -9,8 +9,8 @@ import (
 	gov "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 
 	appmsgs "github.com/dydxprotocol/v4-chain/protocol/app/msgs"
+	"github.com/dydxprotocol/v4-chain/protocol/lib"
 	"github.com/dydxprotocol/v4-chain/protocol/lib/ante"
-	"github.com/dydxprotocol/v4-chain/protocol/lib/maps"
 	testmsgs "github.com/dydxprotocol/v4-chain/protocol/testutil/msgs"
 
 	"github.com/stretchr/testify/require"
@@ -42,7 +42,7 @@ func TestIsNestedMsg_Empty(t *testing.T) {
 }
 
 func TestIsNestedMsg_Invalid(t *testing.T) {
-	allMsgsMinusNested := maps.MergeAllMapsMustHaveDistinctKeys(appmsgs.AllowMsgs, appmsgs.DisallowMsgs)
+	allMsgsMinusNested := lib.MergeAllMapsMustHaveDistinctKeys(appmsgs.AllowMsgs, appmsgs.DisallowMsgs)
 	for key := range appmsgs.NestedMsgSamples {
 		delete(allMsgsMinusNested, key)
 	}
