+++ protocol/lib/ante/unsupported_msgs_test.go
@@ -6,8 +6,8 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	appmsgs "github.com/dydxprotocol/v4-chain/protocol/app/msgs"
+	"github.com/dydxprotocol/v4-chain/protocol/lib"
 	"github.com/dydxprotocol/v4-chain/protocol/lib/ante"
-	"github.com/dydxprotocol/v4-chain/protocol/lib/maps"
 	testmsgs "github.com/dydxprotocol/v4-chain/protocol/testutil/msgs"
 	"github.com/stretchr/testify/require"
 )
@@ -32,7 +32,7 @@ func TestIsUnsupportedMsg_Empty(t *testing.T) {
 }
 
 func TestIsUnsupportedMsg_Invalid(t *testing.T) {
-	allMsgsMinusUnsupported := maps.MergeAllMapsMustHaveDistinctKeys(appmsgs.AllowMsgs, appmsgs.DisallowMsgs)
+	allMsgsMinusUnsupported := lib.MergeAllMapsMustHaveDistinctKeys(appmsgs.AllowMsgs, appmsgs.DisallowMsgs)
 	for key := range appmsgs.UnsupportedMsgSamples {
 		delete(allMsgsMinusUnsupported, key)
 	}
