+++ protocol/lib/collections_test.go
@@ -1,7 +1,6 @@
 package lib_test
 
 import (
-	"math/big"
 	"sort"
 	"testing"
 
@@ -223,4 +222,82 @@ func TestSliceToSet_PanicOnDuplicate(t *testing.T) {
 			lib.SliceToSet(stringSlice)
 		},
 	)
+}
+
+func TestMergeAllMapsWithDistinctKeys(t *testing.T) {
+	tests := map[string]struct {
+		inputMaps []map[string]string
+
+		expectedMap map[string]string
+		expectedErr bool
+	}{
+		"Success: nil input": {
+			inputMaps:   nil,
+			expectedMap: map[string]string{},
+			expectedErr: false,
+		},
+		"Success: single map": {
+			inputMaps: []map[string]string{
+				{"a": "1", "b": "2"},
+			},
+			expectedMap: map[string]string{
+				"a": "1", "b": "2",
+			},
+			expectedErr: false,
+		},
+		"Success: single map, empty": {
+			inputMaps:   []map[string]string{},
+			expectedMap: map[string]string{},
+			expectedErr: false,
+		},
+		"Success: multiple maps, all empty or nil": {
+			inputMaps: []map[string]string{
+				{}, nil,
+			},
+			expectedMap: map[string]string{},
+			expectedErr: false,
+		},
+		"Success: multiple maps, some empty": {
+			inputMaps: []map[string]string{
+				{}, nil, {"a": "1", "b": "2"},
+			},
+			expectedMap: map[string]string{
+				"a": "1", "b": "2",
+			},
+			expectedErr: false,
+		},
+		"Success: multiple maps, no empty": {
+			inputMaps: []map[string]string{
+				{"a": "1", "b": "2"},
+				{"c": "3", "d": "4"},
+			},
+			expectedMap: map[string]string{
+				"a": "1", "b": "2", "c": "3", "d": "4",
+			},
+			expectedErr: false,
+		},
+		"Error: duplicate keys": {
+			inputMaps: []map[string]string{
+				{"a": "1", "b": "2"},
+				{"c": "3", "d": "4"},
+				{"a": "5"}, // duplicate key
+			},
+			expectedMap: nil,
+			expectedErr: true,
+		},
+	}
+
+	for name, tc := range tests {
+		t.Run(name, func(t *testing.T) {
+			if tc.expectedErr {
+				require.PanicsWithValue(
+					t,
+					"duplicate key: a",
+					func() { lib.MergeAllMapsMustHaveDistinctKeys(tc.inputMaps...) })
+			} else {
+				actualMap := lib.MergeAllMapsMustHaveDistinctKeys(tc.inputMaps...)
+				require.Equal(t, tc.expectedMap, actualMap)
+			}
+		})
+	}
 }
